+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- FEAT([199](https://github.com/meateam/api-gateway/pull/199)): added pagination, BREAKING shared files return value
+
 ## [v2.1.0] - 2020-11-1
 
 ### Added
@@ -18,8 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - FEAT([153](https://github.com/meateam/api-gateway/pull/153)): add update feature
 
 - FEAT([162](https://github.com/meateam/api-gateway/pull/162)): add auth startegy for docs
-
-## [Unreleased]
 
 [unreleased]: https://github.com/meateam/api-gateway/compare/master...develop
 [v2.1.0]: https://github.com/meateam/api-gateway/compare/v2.0.0...v2.1.0